requirements/base.txt
Django==2.0.8
django-ckeditor==5.3.1
djangorestframework==3.8.2
